Move protocols into separate .swift files for better code organization and maintainability
